Use Long's for ID; Accept JSON body

+++ src/main/java/dev/submelon/pantry/PantryItemRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PantryItemRepository extends CrudRepository<PantryItem, Integer> {
+public interface PantryItemRepository extends CrudRepository<PantryItem, Long> {
 
     Optional<PantryItem> findByName(String name);
